Use the PluginsController to initialise all plugins
All other plugin controllers must then grab references to the plugins
that it needs from the base plugin controller. This allows references to
be shared across plugin controllers, reducing object creation, but more
importantly, creates only 1 instance of a plugin, so there is always
only one canonical set of data for each plugin.

A consequence of this is that all plugin types must delcare a self-type
annotation on Plugin.