Fix initial text being undoable when editing nodes
There isn't actually a way to set an initial value to the content in the
rich text field beyond an insertion or replacement of the empty text.
But the insertion/replacement is recognised as an action by the undo
manager, and hence is immediately undoable despite there being no edits.

To stop the original text from being undoable, the undo manager is
forced to forget the undo history, leaving the initial
insertion/replacement out of the undo manager.

Reference: https://github.com/TomasMikula/RichTextFX/issues/140