Allow calls to StoryController actions to mark if an action is undoable
Completion events to be propagated on undo/redo, so that plugins can
respond appropriately undo/redo events. Thus, the normal actions of the
StoryController must be used in undo/redo to allow event propagation. At
the same time, changes cannot be sent out on every action call,
otherwise it would result in a feedback loop, where an undo/redo
generates more change events.

To solve that issue, the controller action has a flag to allow the
caller to specify if the action is undoable. This defaults to true, to
not break existing code in the EventDispatcher, and to automatically
handle undo/redo.