Use the provided id to create the node if the id is valid
This is to accomodate undo/redo, and possibly future saving/loading. The
destroy event relies on the existence of the id to destroy the
appropriate node. This id is stored in the undo/redo change. Without
replicating the id on undo/redo, undo/redo actions that remove the node
will not work as it will become unable to find the id to remove.