Add Events API
The events API consists of the Event traits and classes, and the
EventBus object.

The Event traits and classes specify what events are available in for
use, while the EventBus provides a way to send the Event to the rest of
the application. The EventBus also exposes a set of event streams that
track each individual event type, for simplicity.