Add serialisation system
Serialisation consists of 2 parts: from/to objects to/from the AST, and
from/to the AST to/from strings.

Serialisation from objects to the AST is based on the typeclass idiom.

The AST <-> String transformation is handled direcly by a JSON converter
currently. It may be worth converting this to the typeclass idiom to
allow for multiple serialisation backends.