Add Save As ability
The program now differentiates between a Save and a Save As.

On a save, if there is a file loaded, it simply saves back to the same
file, otherwise, it will ask for a file to save to.

On a Save As, it will always ask for a file to save to.