Redesign Rule system
The rules redesign starts from a different angle: instead of considering
rules in isolation, and how those rules are to be associated to the
related objects (and the related typing rules), it is considered from
the angle of how rules are defined.

Thus, conditions and actions define their names (and consequently their
type), and all their associated parameters with names, including the
possible values those parameters may take. These possible values are
noted in ParameterValues.scala.

The redesign creates a much easier way to reason about the rules. A rule
is thus an instance of a definition, with parameter names bound to one
of the specific available values.