Convert generic fact list into specialised fact list
Because of type erasure on the JVM, it is impossible (without reflection
acrobatics) to determine the full type of a list (e.g., whether it is a
list of integers or a list of strings) at runtime. But that type is
needed in order to provide the correct UI for each list type eventually.

In order to have the correct runtime UI, lists have been refactored to
include their specialised type.