Rename find to reflect the object it's finding
Because NodeId and FactId are value types, at runtime they have the same
runtime type (BigInt). `find` is overloaded on FactId and NodeId, but
because they have the same type at runtime, it is impossible to
differentiate them, making it an error to overload on FactId and NodeId.

Also, both methods return an Option, which is a generic type, and hence
due to type erasure on the JVM, Option[Node] and Option[Fact] are the
same, further making it impossible to differentiate them at runtime.