Add functions for converting from the text area data structures to required structures
The text area uses something called `StyleSpan`s to style text.
StyleSpans are contain 2 things: a length that determines how many chars
a StyleSpan covers, and the actual style the the span uses. StyleSpans
are implicitly moved as text is entered into a text area, as the spans
around a particular span changes.

The problem is that we need the actual positions in the text for which a
ruleset applies, so we need to convert from the text span lengths to
actual positions inside the text. This is done by converting cumulative
sums of spans into text positions, since the cumulative sum of text span
lengths before a span is the starting position of that span within the
text, and that starting position added with its length gives the ending
position of the span.