Correct wrong modelling of text rules
Text rules are not simply a single rule attached to some text, but a
list of rules attached to some text. Which means a simple mapping of
indexes to rules will not suffice. Furthermore, text rules have a name
attached to them, which further strengthens the case that a map is not
enough.

To remedy that, a new data structure, the Ruleset, is used to model the
text rules.