Wrap the rules provided to the rule pane in a property
This allows binding the rules to the items of the rules pane, which
allows the UI to update when either the list of rules is changed, or
when items are added/removed from the current list.

This is mostly for the display of text rules, as each text rule has its
own list of rules, which must be swapped out when a text rule is
selected/deselected. Without the binding, swapping out the entire list
does not trigger a redraw, and will leave the list view blank when
adding rules to an empty list