Add reimplementation of story viewer
The reimplementation of the storyviewer uses properties to maintain and
update state whenever possible, to simplify updating, and remove the
need to do manual updating.

Also, link updating is handled via removing all of a node's links to
other nodes, then re-adding them back. This means that the display order
of links is not stable, though hopefully it won't be a problem