Unset old start node when a node is set as a start node
When a node is updated to be a start node, the old start node is set to
no longer be a start node. The updated `update` function makes a note of
whether a node was changed if the node to be updated is set as the start
node.

The undo change now also takes this into account