Simplify code in story controller for handing object creation
Use `copy` instead of manually invoking the constructor. This results in
less things to type into the constructor, and makes it clearer what
changed between the original and the changed objects.

As a side benefit, it also automatically accommodates API breakages in
the unchaged object content, since anything not changed is automatically
copied over into the new object.