Fix #8
When updating rulesets, instead of just updating where each ruleset is
supposed to be in the text, we now also do a diff of the rulesets in the
text and the rulesets of the node. If a ruleset exists in the node, but
not in the text, it is removed from the node.