Change the rules listing to use a TreeView instead of a ListView
By using a TreeView, we can insert a new RuleCell into the list
directly, instead of having to place a TreeView inside a ListView.

The downside of this is that we now have to manage what happens when the
rules list is modified (an issue for deciding what text rules to
display), and we have to add and remove items from the TreeView
manually.

The upside is that we no longer have to deal with resizing issues, and
no longer have scroll bars per rule cell.