Clean up export and story run
Export and has been cleaned up to use the new IoController function to
handle copying resources out of the JAR.

Also, instead of dealing with paths, all file-related executions now use
files directly. The assumption is that we won't run into file path
issues on Windows, but the assumption needs testing.

Running a story has been delegated to the UI, because browser execution
is part of various UI toolkits.