Fix #38
The problem was that asking the undo manager to forget history only
forgets the undo history, and leaves redo history alone. So, to forget
redo history, we need to replace the entire undo manager.

The result is wrapping the undo manager in a property, and expose the
streams of the properties we need via EasyBind flat mappings (for an
effective auto-rebind when the undo manager changes).