Fix null pointer exceptions being thrown on combobox items changing
When the combobox's items change, it sets the value of the selected item
to be null, so we reset the item in the earlier commit that fixes #40 to
reset the selected value if it had one.

But because the value is still being set to null before being reset,
that fires an onAction event, which then tries to update the value in
the paramMap, but the selected item is null, and we try to get some
property of a null object, that throws a NullPointerException.

To fix that, we wrap the possibly null value in an option, and only
update the paramMap if it's not null.