Add keyboard controls for zoom
Partial fix for #42: shortcut +/- to zoom in/out, and shortcut 0 to
reset zoom.

Adding menu items is a bit more complicated because there currently
exists no way for a plugin to hook into the menu and add its own items.
So before menu items can be added, the problem of how to expose points
in which to add menu items need to be solved.