Fix #45
The problem was that `styleRangeAt` was giving the range with respect to
the paragraph the range was in, but the code (and `selectRange`) was
assuming the range was an absolute range with respect to the whole text.
So the 2nd click ended up selecting with respect to the whole text, but
with indexes based on paragraph.

In fact, to be honest, I now have no idea why the first click worked in
the first place.