Fix selection behaviour
The original code forcibly deselected every link and node when any point in the story viewer was clicked, before any other code was run for clicking. This meant that clicking nodes would cause them to be deselected then reselected, which meant that clicking on a node would never be able to deselect it.

This corrects that behaviour by only deselecting nodes and links that do not contain the clicked point, allowing the code that handles node and link selection/deselection to run correctly, allowing deselection via clicking,