Move index updating back to when node content text is not being updated
For copy/paste, we need to update the rulesets in the text whenever text
is pasted, so there is a need to update once the paste is done
(considered as when the text is not updating).

To fix the issue of node updating causing rulesets to disappear, we keep
track of the subscription created by the onChange listener, cancel it
before updating the node data, then reapply the onChange listener once
the text data is updated.