Enable copy/paste of text with its ruleset
Fix #46

Copy/paste of text with style information is done with StyleCodecs in
RichTextFX. A style codec is implemented to copy/paste the ruleset
associated with a style section.

The implementation breaks the abstraction layer between the UI and the
core in order to make use of existing serialisers for the ruleset, and
the JSON serialiser to create a string to store the information. This
was done because reimplementating the serialiser on the UI level is
painful and a lot of work. Future work should look at moving the
object->AST serialisers to the API, and calling a JSON serialiser
manually in the UI, so as to correctly maintain abstraction.