Fix #60
This happens because when redo-ing an action, the undo event dispatcher
relies on the core event dispatcher to do the work, but the core event
dispatcher does not use all the information of the undo event to send a
new change to the undo stream, so the changes being sent end up being
different from the ones that are expected.

To be more concrete, here, when the redo is applied, the core event
dispatcher sends a change with the created node to the undo stream, but
that change is supposed to include the nodes changed as a result of the
node's creation, which the core event dispatcher assumes is empty
because it's create is supposed to be used only in creating pure new
nodes, not re-creating deleted nodes.

To fix it, when nodes are recreated on redo, the core event dispatcher
does not send undo events, instead, the change itself informs the undo
stream of the change being applied, since it's the one with all the
necessary information to re-create a node.