Fix #59
This is a similar problem to #60, where the change is being
inconsistently produced. In this case, the updated node change sees that
the edited node is a start node and incorrectly updates the start node
when the node being edited is already the start node before it was
edited. In order to correctly update the start node, we check if the
node being edited was not already the start node, and only perform start
node updates then.