Fix #63
In all likelihood, this is caused by the filtering for appropriate
values to get the right point to put the arrowhead at, caused by
floating point errors on the 1.0 value. To fix, we simply assume if no
intersection point exists, just take the end point as the place to draw
the arrowhead.