Fix #64
Due to the way changes are implemented, having the core send changes on
undo seems to result in inconsistent change sending. So, to fix the
problem, the core now only sends changes when undo is not being carried
out (like before).

Instead, we now have changes in charge of informing the change stream
that they have been applied. For the nodes anyway. If this pops up in
other places involving undo, we can patch that then.