Handle OS X quit action
Partially fixes #67. Cmd-Q was being sent to the app, but that is separately handled from the main exit handlers.

In the long term, especially once JavaFX can properly handle Mac-isms, we should look into replacing the whole menu bar on the Mac. It's more maintenence work, but given the number of annoying Mac-isms that need to be handled, there's enough differences that it's might not be worth sharing a menu bar across all platforms.