Correct icon sizes
See http://speling.shemnon.com/blog/2013/01/28/adding-icons-with-the-gradle-javafx-plugin/
for details on icon conventions and sizes.

Note that OS X only wants/needs icons of 16x16, 32x32, 128x128, 256x256,
and 512x512, as well as the equivalent double resolutions for HiDPI screens.

I've kept the 2048x2048 icon because that's the best icon for resizing when
necessary.