Remove fact equivalence via id values
Simple fact equivalence via the id values seem to be causing more trouble than it's worth (we want to also deal with changes that are a result of other parts of the fact as well). Might be worth looking at completely removing any manual overrides of equality, to ensure this kind of problems don't happen again. If I'm not wrong, any time we want id equivalence, we check for it manually already anyway.

Fixes #71