Fixes #87
This problem exists because each click of a double click triggers the mouse pressed event. To overcome that, we explicitly force a
select on double click. Note, however, that there is a bit of a delay between the deselection and the forced reselection, so it may look
unselected before the editor opens.