Fix #86
This is happening because the scaled locations of the nodes are being saved, and then when being loaded, the scaled locations are scaled again, causing the positions to change again on load. To fix, we simply save the unscaled locations, instead of the scaled locations, allowing the scaling on load to work correctly.