Fix #81
Recursion is used to go through all the children and sub-children to collapse/expand them. Note that this is not tail recursive, and is effectively a depth-first search, which could be very troublesome to turn into tail recursion. Hopefully, we won't explode the stack depth of the JVM, because the depth of the tree isn't particularly deep.