Add typeclass implementations of new Serialisable and Deserialisable typeclasses
This also entails a change in the file format. The rulesets used to be stored
separately from the text; now they are stored as pairs, where each text may have
an accompanying ruleset. This simplifies processing in the new reader.

The old file format is still readable, but all saves will move to the new file
format.