Switch from EasyBind to Sarah’s ScalaFX Utilities (scalafx-util)
Significantly cleans up syntax for various mappings we need to do, and allows us to better treat Observables as a collection.

The only leftover is the tooltip handling in the fact viewer, because map/flatMap assumes non-null objects, HypeDyn will crash if we switch the select chain in the fact viewer to maps and flatMaps.