Fix #130
Undo seems to be executed synchronously, meaning the undo manager checks if it receives the right change when the undo() in the undo class has finished executing, but most of our actions execute asynchronously. This means the undo manager checks to see if it gets the right change before our system has a chance to perform the change correctly (and consequently emit the right change to the undo stream).

To fix, the CoreEventDispatcher no longer send changes for undo actions to the undo stream, and the change now directly emits undo changes to the undo stream. (This is what the node changes do, which is probably why we’ve never hit this problem with the node undo.)